backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
gunicorn==21.2.0
PyJWT==2.8.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg[binary]==3.1.18
